Conditionally patch Chromium sources with Bromite patch-set

+++ hooks/build_chromium_pre.sh
@@ -7,8 +7,10 @@
 # cd "${CHROMIUM_BUILD_DIR}/src"
 # git am --whitespace=nowarn ${vanadium_dir}/patches/*.patch
 
-echo "applying bromite patches"
-cd "${CHROMIUM_BUILD_DIR}/src"
-rm -rf .git/rebase-apply/
-# Apply the patches in the given order
-while IFS= read -r line; do git am --whitespace=nowarn "${BROMITE_DIR}/build/patches/$line"; done < "${BROMITE_DIR}/build/bromite_patches_list.txt"
+if [ "${APPLY_BROMITE_PATCHES}" = true ]; then
+    echo "applying bromite patches"
+    cd "${CHROMIUM_BUILD_DIR}/src"
+    rm -rf .git/rebase-apply/
+    # Apply the patches in the given order
+    while IFS= read -r line; do git am --whitespace=nowarn "${BROMITE_DIR}/build/patches/$line"; done < "${BROMITE_DIR}/build/bromite_patches_list.txt"
+fi
